Refactor BaseController and enhance telemetry handling; update LoRa configuration and LTE connection management

+++ src/model/Telemetry.h
@@ -18,7 +18,7 @@ struct Telemetry {
   int baseBatteryPercent;
   bool isAlerted;
   int alertType;
-  int nodeVbus;
+  float nodeVbus;
   int nodeHasBattery;
 };
